+++ CHANGELOG.md
@@ -8,6 +8,7 @@ See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/)
 ### Added
 
 - upcloud_loadbalancer_frontend_rule: added `matching_condition` field.
+- provider: `ProviderConfigure()` and `NewWithUserAgent()` to allow setting user agent
 
 ### Fixed